identicon: generate missing icons at compile time; expose `generateAssetIcon` api for other applications
